User commit

+++ src/main/java/com/example/demo/controller/ValidationController.java
@@ -22,7 +22,7 @@ public class ValidationController{
         return service.postdata(val);
     }
     @GetMapping("/valfind/{id}")
-    public ValidationEntity find(@PathVariable int id){
+    public ValidationEntity find(@PathVariable Long id){
         return service.getdata(id);
     }
 }
